pyproject.toml
[tool.poetry]
name = "solana-bot"
version = "0.1.0"
description = ""
authors = ["Abderrahmen Trabelsi"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.2.2"
scipy = "^1.15.1"
asyncpg = "^0.30.0"
loguru = "^0.7.3"
backtrader = "^1.9.78.123"
pandas = "^2.2.3"
pynacl = "^1.5.0"
base58 = "^2.1.1"
python-dotenv = "^1.0.1"
aiohttp = "^3.11.12"
peewee = "^3.17.9"
peewee-async = "^1.0.0"
websockets = "^14.2"
solana = "^0.36.5"
ta-lib = {path = "ta_lib-0.6.3-cp311-cp311-win_amd64.whl"}
solders = "^0.25.0"
winuvloop = "^0.2.0"
aiopg = "^1.4.0"


[tool.poetry.group.dev.dependencies]
poetry = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "src.main:run_bot"
keygen = "src.utils.keygen:main"
